+++ packages/core/src/slate/editor/TEditor.ts
@@ -2,10 +2,17 @@ import { Editor } from 'slate';
 import { UnknownObject } from '../../common/types/utility/AnyObject';
 import { Modify } from '../../common/types/utility/types';
 import { EElement, EElementOrText, TElement } from '../element/TElement';
-import { ENodeEntry } from '../node/TNodeEntry';
+import { TDescendant } from '../node/TDescendant';
+import { ENode, TNode } from '../node/TNode';
+import { TNodeEntry } from '../node/TNodeEntry';
 import { TOperation } from '../types/TOperation';
 
 export type Value = TElement[];
+
+/**
+ * A helper type for getting the value of an editor.
+ */
+export type ValueOf<E extends TEditor<Value>> = E['children'];
 
 export type TEditor<V extends Value> = Modify<
   Editor,
@@ -15,22 +22,40 @@ export type TEditor<V extends Value> = Modify<
     marks: Record<string, any> | null;
 
     // Schema-specific node behaviors.
-    isInline: <EV extends Value>(element: EElement<EV>) => boolean;
-    isVoid: <EV extends Value>(element: EElement<EV>) => boolean;
-    normalizeNode: <EV extends Value>(entry: ENodeEntry<EV>) => void;
+    isInline: <N extends TElement>(element: N) => boolean;
+    isVoid: <N extends TElement>(element: N) => boolean;
+    normalizeNode: <N extends TNode>(entry: TNodeEntry<N>) => void;
 
     // Overrideable core actions.
-    apply: (operation: TOperation) => void;
-    getFragment: <EV extends Value>() => EElementOrText<EV>[];
-    insertFragment: <EV extends Value>(fragment: EElementOrText<EV>[]) => void;
-    insertNode: <EV extends Value>(
-      node: EElementOrText<EV> | EElementOrText<EV>[]
-    ) => void;
+    apply: <N extends TDescendant>(operation: TOperation<N>) => void;
+    getFragment: <N extends TDescendant>() => N[];
+    insertFragment: <N extends TDescendant>(fragment: N[]) => void;
+    insertNode: <N extends TDescendant>(node: N | N[]) => void;
   }
 > &
   UnknownObject;
 
 /**
- * A helper type for getting the value of an editor.
+ * Get editor with typed methods and operations.
+ * Note that it can't be used as a parameter of type TEditor.
  */
-export type ValueOf<E extends TEditor<Value>> = E['children'];
+export const getTEditor = <V extends Value, E extends TEditor<V> = TEditor<V>>(
+  editor: E
+) =>
+  editor as Modify<
+    E,
+    {
+      operations: TOperation<EElementOrText<V>>[];
+
+      // Schema-specific node behaviors.
+      isInline: (element: EElement<V>) => boolean;
+      isVoid: (element: EElement<V>) => boolean;
+      normalizeNode: (entry: TNodeEntry<ENode<V>>) => void;
+
+      // Overrideable core actions.
+      apply: (operation: TOperation<EElementOrText<V>>) => void;
+      getFragment: () => EElementOrText<V>[];
+      insertFragment: (fragment: EElementOrText<V>[]) => void;
+      insertNode: (node: EElementOrText<V> | EElementOrText<V>[]) => void;
+    }
+  >;